+++ src/main/java/dev/lambdaurora/lambdynlights/mixin/DebugScreenOverlayMixin.java
@@ -10,6 +10,7 @@
 package dev.lambdaurora.lambdynlights.mixin;
 
 import dev.lambdaurora.lambdynlights.LambDynLights;
+import dev.lambdaurora.lambdynlights.LambDynLightsConstants;
 import dev.lambdaurora.lambdynlights.engine.DynamicLightingEngine;
 import net.minecraft.TextFormatting;
 import net.minecraft.client.gui.components.DebugScreenOverlay;
@@ -24,7 +25,7 @@ import java.util.List;
  * Adds a debug string for dynamic light sources tracking and updates.
  *
  * @author LambdAurora
- * @version 3.2.0
+ * @version 3.2.2
  * @since 1.3.2
  */
 @Mixin(DebugScreenOverlay.class)
@@ -49,5 +50,9 @@ public class DebugScreenOverlayMixin {
 
 		builder.append(')');
 		list.add(builder.toString());
+
+		if (LambDynLightsConstants.isDevMode()) {
+			list.add(TextFormatting.RED + LambDynLightsConstants.DEV_MODE_OVERLAY_TEXT);
+		}
 	}
 }